add test - password exception test
